# Implementing fast Latin1 to UTF functions for ARM.