This PR provides fast function (with ARM NEON) for Latin 1 transcoding to Unicode formats.

Here are some benchmarks on an M2 processor with LLVM 14. I use the french.latin1 file from the wikipedia mars dataset.

The performance is good.

## to UTF-8

```
$ ./benchmarks/benchmark -P convert_latin1_to_utf8 -F ../unicode_lipsum/wikipedia_mars/french.latin1.txt -I 3000
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
Using ICU version 73.2
Using iconv version 267
testcases: 1
input detected as unknown
current system detected as arm64
===========================
convert_latin1_to_utf8+arm64, input size: 432305, iterations: 3000, dataset: ../unicode_lipsum/wikipedia_mars/french.latin1.txt
   6.963 GB/s (3.4 %)    6.963 Gc/s     1.00 byte/char 
convert_latin1_to_utf8+iconv, input size: 432305, iterations: 3000, dataset: ../unicode_lipsum/wikipedia_mars/french.latin1.txt
   0.322 GB/s (6.0 %)    0.322 Gc/s     1.00 byte/char 
convert_latin1_to_utf8+icu, input size: 432305, iterations: 3000, dataset: ../unicode_lipsum/wikipedia_mars/french.latin1.txt
   1.328 GB/s (4.7 %)    1.328 Gc/s     1.00 byte/char 
```

## to UTF-16

```
$ ./benchmarks/benchmark -P convert_latin1_to_utf16 -F ../unicode_lipsum/wikipedia_mars/french.latin1.txt -I 3000
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
Using ICU version 73.2
Using iconv version 267
testcases: 1
input detected as unknown
current system detected as arm64
===========================
convert_latin1_to_utf16+arm64, input size: 432305, iterations: 3000, dataset: ../unicode_lipsum/wikipedia_mars/french.latin1.txt
  35.053 GB/s (7.6 %)   35.053 Gc/s     1.00 byte/char 
convert_latin1_to_utf16+iconv, input size: 432305, iterations: 3000, dataset: ../unicode_lipsum/wikipedia_mars/french.latin1.txt
   0.343 GB/s (4.3 %)    0.343 Gc/s     1.00 byte/char 
convert_latin1_to_utf16+icu, input size: 432305, iterations: 3000, dataset: ../unicode_lipsum/wikipedia_mars/french.latin1.txt
   6.389 GB/s (4.5 %)    6.389 Gc/s     1.00 byte/char 
```

## to UTF-32

```
$ ./benchmarks/benchmark -P convert_latin1_to_utf32 -F ../unicode_lipsum/wikipedia_mars/french.latin1.txt -I 3000
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
Using ICU version 73.2
Using iconv version 267
testcases: 1
input detected as unknown
current system detected as arm64
===========================
convert_latin1_to_utf32+arm64, input size: 432305, iterations: 3000, dataset: ../unicode_lipsum/wikipedia_mars/french.latin1.txt
  18.235 GB/s (6.9 %)   18.235 Gc/s     1.00 byte/char 
convert_latin1_to_utf32+iconv, input size: 432305, iterations: 3000, dataset: ../unicode_lipsum/wikipedia_mars/french.latin1.txt
   0.380 GB/s (4.7 %)    0.380 Gc/s     1.00 byte/char 
convert_latin1_to_utf32+icu, input size: 432305, iterations: 3000, dataset: ../unicode_lipsum/wikipedia_mars/french.latin1.txt
   0.292 GB/s (4.8 %)    0.292 Gc/s     1.00 byte/char 
```

Should fix... https://github.com/simdutf/simdutf/issues/276 https://github.com/simdutf/simdutf/issues/278 https://github.com/simdutf/simdutf/issues/280